+++ lib/data/extension/copy_directory.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:genshin_mod_manager/extension/pathops.dart';
+import 'package:genshin_mod_manager/data/extension/pathops.dart';
 
 extension CopyDirectory on Directory {
   void copyToPath(String dest) {
