+++ lib/domain/service/app_state_service.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:genshin_mod_manager/extension/pathops.dart';
+import 'package:genshin_mod_manager/data/extension/pathops.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class AppStateService with ChangeNotifier {
