+++ lib/ui/route/home_shell.dart
@@ -8,16 +8,16 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
-import 'package:genshin_mod_manager/base/appbar.dart';
-import 'package:genshin_mod_manager/extension/pathops.dart';
-import 'package:genshin_mod_manager/io/fsops.dart';
-import 'package:genshin_mod_manager/service/app_state_service.dart';
-import 'package:genshin_mod_manager/service/folder_observer_service.dart';
-import 'package:genshin_mod_manager/service/preset_service.dart';
-import 'package:genshin_mod_manager/third_party/fluent_ui/auto_suggest_box.dart';
-import 'package:genshin_mod_manager/third_party/fluent_ui/red_filled_button.dart';
-import 'package:genshin_mod_manager/widget/category_drop_target.dart';
-import 'package:genshin_mod_manager/widget/preset_control.dart';
+import 'package:genshin_mod_manager/data/extension/pathops.dart';
+import 'package:genshin_mod_manager/data/io/fsops.dart';
+import 'package:genshin_mod_manager/domain/service/app_state_service.dart';
+import 'package:genshin_mod_manager/domain/service/folder_observer_service.dart';
+import 'package:genshin_mod_manager/domain/service/preset_service.dart';
+import 'package:genshin_mod_manager/ui/widget/base/appbar.dart';
+import 'package:genshin_mod_manager/ui/widget/category_drop_target.dart';
+import 'package:genshin_mod_manager/ui/widget/preset_control.dart';
+import 'package:genshin_mod_manager/ui/widget/third_party/fluent_ui/auto_suggest_box.dart';
+import 'package:genshin_mod_manager/ui/widget/third_party/fluent_ui/red_filled_button.dart';
 import 'package:go_router/go_router.dart';
 import 'package:http/http.dart' as http;
 import 'package:logger/logger.dart';
