+++ lib/ui/route/setting.dart
@@ -1,7 +1,7 @@
 import 'package:filepicker_windows/filepicker_windows.dart';
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:genshin_mod_manager/service/app_state_service.dart';
-import 'package:genshin_mod_manager/third_party/no_deref_file_opener.dart';
+import 'package:genshin_mod_manager/domain/service/app_state_service.dart';
+import 'package:genshin_mod_manager/ui/widget/third_party/flutter/no_deref_file_opener.dart';
 import 'package:go_router/go_router.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:provider/provider.dart';
