+++ lib/ui/widget/chara_mod_card.dart
@@ -4,14 +4,14 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:genshin_mod_manager/extension/pathops.dart';
-import 'package:genshin_mod_manager/io/fsops.dart';
-import 'package:genshin_mod_manager/route/nahida_store.dart';
-import 'package:genshin_mod_manager/service/folder_observer_service.dart';
-import 'package:genshin_mod_manager/third_party/fluent_ui/red_filled_button.dart';
-import 'package:genshin_mod_manager/upstream/akasha.dart';
-import 'package:genshin_mod_manager/widget/editor_text.dart';
-import 'package:genshin_mod_manager/widget/toggleable.dart';
+import 'package:genshin_mod_manager/data/extension/pathops.dart';
+import 'package:genshin_mod_manager/data/io/fsops.dart';
+import 'package:genshin_mod_manager/data/upstream/akasha.dart';
+import 'package:genshin_mod_manager/domain/service/folder_observer_service.dart';
+import 'package:genshin_mod_manager/ui/route/nahida_store.dart';
+import 'package:genshin_mod_manager/ui/widget/editor_text.dart';
+import 'package:genshin_mod_manager/ui/widget/third_party/fluent_ui/red_filled_button.dart';
+import 'package:genshin_mod_manager/ui/widget/toggleable.dart';
 import 'package:logger/logger.dart';
 import 'package:pasteboard/pasteboard.dart';
 import 'package:provider/provider.dart';
