+++ test/welcome_route_test.dart
@@ -1,6 +1,6 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:genshin_mod_manager/route/welcome.dart';
+import 'package:genshin_mod_manager/ui/route/welcome.dart';
 import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
 
 import 'mock/mock_url_launcher_platform.dart';
